+++ config/cfg.go
@@ -51,17 +51,19 @@ func DefaultConfig() *Config {
 
 		ProtectMode: true,
 
-		HTTPListen: "localhost:9529",
 		HTTPAPI: &dkhttp.APIConfig{
 			RUMOriginIPHeader: "X-Forwarded-For",
+			Listen:            "localhost:9529",
 		},
 
 		IOCacheCount: 1024,
 
-		LogLevel:  "info",
-		LogRotate: 32,
-		Log:       filepath.Join("/var/log/datakit", "log"),
-		GinLog:    filepath.Join("/var/log/datakit", "gin.log"),
+		Logging: &LoggerCfg{
+			Level:  "info",
+			Rotate: 32,
+			Log:    filepath.Join("/var/log/datakit", "log"),
+			GinLog: filepath.Join("/var/log/datakit", "gin.log"),
+		},
 
 		BlackList: []*inputHostList{
 			&inputHostList{Hosts: []string{}, Inputs: []string{}},
@@ -74,8 +76,8 @@ func DefaultConfig() *Config {
 
 	// windows 下，日志继续跟 datakit 放在一起
 	if runtime.GOOS == "windows" {
-		c.Log = filepath.Join(datakit.InstallDir, "log")
-		c.GinLog = filepath.Join(datakit.InstallDir, "gin.log")
+		c.Logging.Log = filepath.Join(datakit.InstallDir, "log")
+		c.Logging.GinLog = filepath.Join(datakit.InstallDir, "gin.log")
 	}
 
 	return c
@@ -87,25 +89,37 @@ type Cgroup struct {
 	CPUMin float64 `toml:"cpu_min"`
 }
 
+type LoggerCfg struct {
+	Log          string `toml:"log"`
+	GinLog       string `toml:"gin_log"`
+	Level        string `toml:"level"`
+	DisableColor bool   `toml:"disable_color"`
+	Rotate       int    `toml:"rotate,omitzero"`
+}
+
 type Config struct {
-	Namespace      string `toml:"namespace"`
 	UUID           string `toml:"-"`
-	Hostname       string `toml:"-"`
 	UUIDDeprecated string `toml:"uuid,omitempty"` // deprecated
 
-	Name    string              `toml:"name,omitempty"`
+	Name      string `toml:"name,omitempty"`
+	Hostname  string `toml:"-"`
+	Namespace string `toml:"namespace"`
+
 	DataWay *dataway.DataWayCfg `toml:"dataway,omitempty"`
 
-	HTTPBindDeprecated string `toml:"http_server_addr,omitempty"`
-	HTTPListen         string `toml:"http_listen,omitempty"`
+	// http config: TODO: merge into APIConfig
+	HTTPBindDeprecated       string            `toml:"http_server_addr,omitempty"`
+	HTTPListenDeprecated     string            `toml:"http_listen,omitempty"`
+	Disable404PageDeprecated bool              `toml:"disable_404page,omitempty"`
+	HTTPAPI                  *dkhttp.APIConfig `toml:"http_api"`
 
-	HTTPAPI *dkhttp.APIConfig `toml:"http_api"`
+	// logging config
+	LogDeprecated       string     `toml:"log,omitempty"`
+	LogLevelDeprecated  string     `toml:"log_level,omitempty"`
+	LogRotateDeprecated int        `toml:"log_rotate,omitzero"`
+	GinLogDeprecated    string     `toml:"gin_log,omitempty"`
+	Logging             *LoggerCfg `toml:"logging"`
 
-	Log       string `toml:"log"`
-	LogLevel  string `toml:"log_level"`
-	LogRotate int    `toml:"log_rotate,omitempty"`
-
-	GinLog       string            `toml:"gin_log"`
 	GlobalTags   map[string]string `toml:"global_tags"`
 	Environments map[string]string `toml:"environments"`
 
@@ -124,7 +138,6 @@ type Config struct {
 	Cgroup    *Cgroup          `toml:"cgroup"`
 
 	EnableElection bool           `toml:"enable_election"`
-	Disable404Page bool           `toml:"disable_404page"`
 	IOCacheCount   int64          `toml:"io_cache_count"`
 	Tracer         *tracer.Tracer `toml:"tracer,omitempty"`
 
@@ -297,18 +310,44 @@ func (c *Config) setupGlobalTags() error {
 	return nil
 }
 
-func (c *Config) ApplyMainConfig() error {
+func (c *Config) setLogging() {
 
 	// set global log root
-	l.Infof("set log to %s", c.Log)
-	if c.Log != "stdout" || c.Log == "" {
-		logger.MaxSize = c.LogRotate
-		if err := logger.SetGlobalRootLogger(c.Log, c.LogLevel, logger.OPT_DEFAULT); err != nil {
+	if c.Logging.Log != "stdout" || c.Logging.Log != "" { // set log to disk file
+
+		l.Infof("set log to %s", c.Logging.Log)
+
+		if c.Logging.Rotate > 0 {
+			logger.MaxSize = c.Logging.Rotate
+		}
+
+		if err := logger.InitRoot(&logger.Option{
+			Path:  c.Logging.Log,
+			Level: c.Logging.Level,
+			Flags: logger.OPT_DEFAULT}); err != nil {
 			l.Errorf("set root log faile: %s", err.Error())
 		}
 	} else {
-		logger.SetStdoutRootLogger(c.LogLevel, logger.OPT_DEFAULT|logger.OPT_STDOUT)
+
+		l.Info("set log to stdout, rotate disabled")
+
+		optflags := (logger.OPT_DEFAULT | logger.OPT_STDOUT)
+		if !c.Logging.DisableColor {
+			optflags |= logger.OPT_COLOR
+		}
+
+		if err := logger.InitRoot(
+			&logger.Option{
+				Level: c.Logging.Level,
+				Flags: optflags}); err != nil {
+			l.Errorf("set root log faile: %s", err.Error())
+		}
 	}
+}
+
+func (c *Config) ApplyMainConfig() error {
+
+	c.setLogging()
 
 	l = logger.SLogger("config")
 
@@ -444,7 +483,7 @@ func (c *Config) LoadEnvs() error {
 	}
 
 	if v := datakit.GetEnv("ENV_DISABLE_404PAGE"); v != "" {
-		c.Disable404Page = true
+		c.HTTPAPI.Disable404Page = true
 	}
 
 	if v := datakit.GetEnv("ENV_GLOBAL_TAGS"); v != "" {
@@ -452,7 +491,11 @@ func (c *Config) LoadEnvs() error {
 	}
 
 	if v := datakit.GetEnv("ENV_LOG_LEVEL"); v != "" {
-		c.LogLevel = v
+		c.Logging.Level = v
+	}
+
+	if v := datakit.GetEnv("ENV_LOG"); v != "" {
+		c.Logging.Log = v
 	}
 
 	// 多个 dataway 支持 ',' 分割
@@ -473,7 +516,7 @@ func (c *Config) LoadEnvs() error {
 	}
 
 	if v := datakit.GetEnv("ENV_HTTP_LISTEN"); v != "" {
-		c.HTTPListen = v
+		c.HTTPAPI.Listen = v
 	}
 
 	if v := datakit.GetEnv("ENV_RUM_ORIGIN_IP_HEADER"); v != "" {